backend/pyproject.toml
[project]
name = "video-extract-pro-backend"
version = "1.0.0"
description = "AI Video Analysis Backend for Video Extract Pro"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=2.0.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",

    # Task queue
    "celery[redis]>=5.3.6",
    "redis>=5.0.1",

    # ML/AI - Transcription
    "faster-whisper>=1.0.1",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",

    # ML/AI - Speaker Diarization
    "pyannote.audio>=3.1.1",

    # ML/AI - Visual Analysis
    "ultralytics>=8.1.0",
    "opencv-python>=4.9.0.80",

    # ML/AI - LLM
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "sentence-transformers>=2.3.1",

    # Video processing
    "yt-dlp>=2024.1.1",
    "moviepy>=1.0.3",
    "ffmpeg-python>=0.2.0",

    # Utilities
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
